last feedback fixes. ctrl deselect, seperate FolderView.cpp
